fix(auth): SSO failed to get token due to missing 'refreshToken'
Problem:

Error users were seeing was:

```
Error: SSO cache data unexpectedly missing props: ["refreshToken"]
```

This was due to a check in a previos PR that assumed the refreshToken
was always present in SSO cache.

Solution:

It looks like the refreshToken does not exist for all cases, so some research needs
to be done.

But for now this reverts the check that the refreshToken exists, but it keeps the validation
that the accessToken exists since that is always guaranteed.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>