ci: GHA jobs re-trigger too frequently
Problem:
Since 6dcbfc705d4df, the `lint-commits` job runs whenever the PR
*description* is edited, which then re-runs all the other GHA jobs. This
is because it listens to the `edited` event.

Solution:
There is no way to avoid rerunning all the CI jobs if they are dependent
on `lint-commits`. Instead, revert 6dcbfc705d4df and document that
re-opening the PR will force CI to re-run, if needed, after fixing a PR
title.