fix(auth): debounce getToken() function
Problem:

The Identity team noticed a large spike in token refreshes for specific users.
One user would trigger refresh over 50 times within a few seconds.

Solution:

The telemetry showed that `getChatAuthState()` was being called many times in a
short period. This eventually triggered the token refresh logic many times, if the token was
expired.

The solution is to add a debounce to `getToken()` which calls the refresh logic.
- `debounce()` only accepts functions without any args, the refresh logic requires args
- `getToken()` will also load from disk is the token is not expired, so debouncing here
  saves disk I/O as well.

The current debounce interval is 100 milliseconds, which based on telemetry should be enough to capture the barrage of calls.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>