fix(amazonq): skip including deleted files for FeatureDev context #6312
## Problem
When zipping context for /dev, customer build processes (file watchers,
etc.) may delete build artifacts we’ve already enumerated but have not
added to the archive. As a best practice, customers should `.gitignore`
these types of files, but in the event they don't, this has the
potential to block /dev from running.

## Solution
Skip affected files, which are not found when adding to zip context for
Feature Dev.