refactor(util): merge withRetries and waitUntil
Problem:

We had 2 different implementation of similar code, between
withRetries and waitUntil.

We couldn't easily merge them though since behavior was slightly different
when it came to expected return types, and since one threw and the other
did not.

- waitUntil returns undefined if it never properly resolves
- withRetries throws if it never properly resolves

Solution:

As a solution we keep both methods but they share the same underlying
code.

- waitUntil now has a backoff field to increase the delta between intervals
  - withRetries also has this
- withRetries will retry even on a thrown exeception
  - on timeout it will throw the last exception it encountered

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>