feat(lambda): user can open CWL Live Tail from Lambda function #6423
Allow to "Tail Logs" from Lambda functions, which will get the
configured Log Group for a particular function and start a Live Tail
session in that Log Group.

Some code was removed, because with the recently updated version of the
SDK v2, we can take the configured log group directly from the
function's configuration and we don't need that extra call using SDK V3
that was being made inside `deployedNode.ts`

## Problem
To start a CloudWatch Live Tail session, a user needs to know their Log
Group name and find it in the AWS explorer. For Lambda users, that might
not be easy to find or well known all the time

## Solution
We help the Live Tail experience for Lambda customers by allowing them
to start a Live Tail session directly from their function in the AWS
explorer and in the Application Builder section for deployed functions,
without having to search for their Log Group explicitly.