feat(q-dev): /dev can be configured to run code, tests #6453
## Problem
Customers cannot configure whether their generated code should be
run and tested for quality assurance before being provided to them.

## Solution
Enable customers to configure whether to run code and
tests with /dev per project. On the first run of /dev for a project, if
a devfile.yaml exists in the root of the project, the customer is
prompted to know if they want to use the commands in the file to test
the code we generate. The feature additionally allows customers to
generate a devfile if one doesn't exist. Their choice is persisted and
configurable in the Amazon Q settings.

Co-authored-by: wilson <wlsun@amazon.com>
Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>
Co-authored-by: Karan Ahluwalia <ahlkaran@amazon.com>
Co-authored-by: Hamed Soleimani <hamed1soleimani@gmail.com>
Co-authored-by: Josh Pinkney <103940141+jpinkney-aws@users.noreply.github.com>