test(flaky): avoid relying on race condition for tested behavior #6452
## Problem
https://github.com/aws/aws-toolkit-vscode/issues/6451

This rename test relies on a specific result of a race condition for the
expected result.
- The test is checking for a telemetry result that is only emitted when
`fs.exists` takes more than 1 attempt to resolve to true.
- Therefore, it wants the first `fs.exists` check to fail, then a
subsequent one to succeed.
- It does this by not awaiting the result, and then writing the file to
be renamed.

Usually this is fine, but it is possible that the write (`toFile`)
happens before the read (`fs.exists`) since neither is awaited. This
behavior leads to a flaky test as described in the issue.

## Solution
- use a stub to force the first call to `fs.exists` to fail. 
- allow all other calls to "go through" to the original function.