fix(amazonq): right click options dont send context if chat not loaded
Problem:

When a user starts up VSC, hasn't opened Q chat yet, then highlights text and
does a right click option like "send to prompt" THEN the whole prompt fails to
make it to chat.

This is because there is a race condition between the message being sent to the webview,
and the webview completed loading and ready to recieve messages.

Solution:

In the appToMessage publisher, verify that the chat is ready to recieve messages based
on if the webview sent the chat ui ready message. Otherwise wait until we get it,
and then send the message.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>