feat(stepfunctions): previewStateMachine opens in Workflow Studio #6511
## Problem
The preview state machine command (used in an icon button and a code
lens button) consists of a read-only graph component which provides a
simple graph representation of an ASL definition. It does not include
any other information such as state types or resources.

## Solution
Replace the graph visualization in the previewStateMachine command with
the Workflow Studio graph and editor. Workflow Studio is editable and
includes icons and descriptions representing the state types and
resources. Users will now be able to edit the state definition in the
graph visualization directly.