fix(build): webviews sometimes did not load during development
Problem:

In our development build we spawn a webpack server that serves the Vue files
used by our webviews. It is served on port 8080, but sometimes an existing
process is already using it, and this causes the webpack server to use the next
available port. But because we assume it will use 8080 it silently breaks.

Solution:

Explicitly require 8080 in the webpack server cli command, also create a custom
script to catch the error early and explain how to resolve it.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>