fix(build): webviews sometimes did not load during development (#6556)
## Problem:

In our development build we spawn a webpack server that serves the Vue
files used by our webviews. It is served on port 8080, but sometimes an
existing process is already using it, and this causes the webpack server
to use the next available port. But because we assume it will use 8080
it silently breaks, and when developing the Q webview would not load as
expected.

## Solution:

Explicitly require 8080 in the webpack server cli command, also create a
custom script to catch the error early and explain how to resolve it.

Now in the build tasks if 8080 is in use the dev can see the failed task
with the solution explicitly stated.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>