ci(jscpd): add hint message for rebasing on false negatives #6564
## Problem
The duplicate code detection can fail when the current branch is not
rebased with the most recent version of master. This is a fairly common
false-negative. To avoid this, the PR must be kept up to date with its
target branch. However, the tool doesn't provide any hints/indication
this is the case.

## Solution
- add a hint message on failure that suggests rebasing.