refactor(amazonq): revert pre-fetch next recommendations for inline completion (#6567)
This reverts commit 50c91210736c1226e8bcb4c1af10fe29fe202ee7.

## Problem
In the previous change, we applied prefect to every responses. However,
this introduced issues with inline completion, causing it to behave
inconsistently and produce unexpected completions in the next response.

## Solution
To mitigate this, this PR rolls back the responses prefect change to
restore stable inline completion behavior.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.