fix(appbuilder): pass in the negative version of --use-container when using build quickpick (#6603)
## Problem
When users use appbuilder to build their lambda functions, they choose
between using their samconfig file or manually selecting the build
parameters/flags. The problem is that when the user selects build flags
and intentionally doesn't select the ```--use-container``` flag, the
command will still be run with --use-container if the samconfig file has
```use_container``` is set to true.

## Solution
Whenever the user manually selects the build flags and doesn't select
```--use-container```, we add the negative version of
```--use-container```, which is ```--no-use-container```. This serves as
an override if the samconfig file has ```--use-container``` set to true.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.