telemetry(auth): emit metric when Auth webview loaded
Problem:

The inital problem is that when we focused chat, it returned succeeded even though it didn't
have any idea if the UI was actually rendered to the user.

Solution:

This new solution will have the UI emit a telemetry metric once per session that indicates
when the UI has loaded. It will also distinguish between the login and reauth page.

The metric is `webview_load` and the `webviewName` field will distinguish between login and reauth

NOTE:

When Q chat itself has its UI ready we already emit a `webview_load` metric. So use all of these metrics
to determine what the user actually saw

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>