check if existing connections caused isFirstUse discrepancy
Before, on a users first use, it was technically possible for Q to indicate it was NOT the first use
if somehow there were existing connections detected.

A previous commit removed that case. But we want to now emit
telemetry if we detected that we WOULD have run in to that case when not
expected.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>