feat(lambda): download serverless land patterns in IDE (#6612)
## Problem
This pull request introduces the functionality to allow users to
download the selected pattern from the Serverless Land QuickPick flow
into a directory of their choice.

## Solution
The key changes made are as follows:
1. The `metadata.json` file was updated to accommodate the addition of
an `asset_name` field for each IaC and Runtime pair.
2. Consequently, the way the metadata is extracted in the
`metadataManager.ts` file had to be modified to account for these
changes in the `metadata.json` file.
3. To enable users to navigate back to the previous QuickPick options, a
switch functionality was implemented in the `wizard.ts` file. This helps
maintain the state of the current QuickPick and directs the user to the
previous QuickPick.
---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Vandita Patidar <vanditap@amazon.com>