fix(ec2): properly paginate instances in quickPick.  (#6682)
## Problem
Follow up to discussion:
https://github.com/aws/aws-toolkit-vscode/pull/6672#discussion_r1972045094

Currently the ec2 module does make paginated requests to the SDK
endpoint. However, in order to show these results into the UI, we
resolve them in entirety.
For the quick picks this is done here:
[https://github.com/aws/aws-toolkit-vscode/blob/4e568970e8d8646cf8e590d39d8891d68cebde84/packages/core/src/awsService/ec2/prompter.ts#L65
](https://github.com/aws/aws-toolkit-vscode/blob/4e568970e8d8646cf8e590d39d8891d68cebde84/packages/core/src/awsService/ec2/prompter.ts#L65)
For the explorer this is done here:
[https://github.com/aws/aws-toolkit-vscode/blob/4e568970e8d8646cf8e590d39d8891d68cebde84/packages/core/src/awsService/ec2/explorer/ec2ParentNode.ts#L52
](https://github.com/aws/aws-toolkit-vscode/blob/4e568970e8d8646cf8e590d39d8891d68cebde84/packages/core/src/awsService/ec2/explorer/ec2ParentNode.ts#L52)

## Solution
- Refactor the EC2 client to process the instances within the pages they
come in (i.e. not flattening them). This allows passing an
`AsyncCollection` of pages to the quickPick.
- Avoid flattening results in `makePaginatedRequest` by default.  
- Refactor client builder to be non-async, this causes some trickle-down
refactoring, but allows `makePaginatedRequest` to not return a promise.
- Simplify instance processing to make use of type predicates. 
- Remove mocks from `Ec2Prompter` tests, and refactor to test pagination
behavior.

## Alternative Solution 
- Repack the instances into artificial pages (i.e. a batched iterator).
This is a little hacky since it involves us unpacking and repacking the
instances. However, it does involve significantly less code changes.

## Verification
In addition to updating the tests, we can add a logging statement when
the quickPick loads an item from the `AsyncIterable`. This screenshot
shows that loading two items was successful and triggered the
`AsyncIterable` twice, as expected.
<img width="1045" alt="image"
src="https://github.com/user-attachments/assets/8303f909-17e9-4fc3-a9fa-b0ec311eb1bc"
/>


## Future Work
- Extend this work to the explorer. There is already some general work
done in
https://github.com/aws/aws-toolkit-vscode/blob/df810aa4264ba14dac488697d539011a732c7eff/packages/core/src/awsexplorer/childNodeLoader.ts,
but it doesn't support working with `AsyncCollection`, only making the
requests directly. We could do something hacky, like wrap the
`AsyncCollection` in a fake request method, or try to refactor the
utility itself.
---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.