fix(stepfunctions): Prevent opening Workflow Studio for invalid JSON/YAML; add save and deploy message handler (#6693)
## Problem
1. Deploy state machine workflow is not very visible to the customers
2. Currently when user is opening WFS with custom action in the side
panel and has invalid JSON/YAML, the side panel opens and immediately
closes (since WFS does not support invalid JSON/YAML). This experience
does not look smooth.

## Solution
1. With Workflow Studio integration, we'll add `Save & Deploy` button to
the UI which will save file, close WFS editor and start the deployment
flow. Adding a message handler for this action
2. Adding a check for invalid JSON/YAML before attempting to open WFS in
the side panel. If file is invalid, the panel will not be opened and the
user will see a notification
---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Vlad Nikolaenko <vladnik@amazon.com>