feat(amazonq): generate scanName from workspace config #6679
## Problem

All code scans are run with a scanName as a UUID which makes it
impossible to track the bug lifecycle of projects. We are not leveraging
the bug tracking feature which allows grouping together different scan
runs of the same project by passing in the same scanName between runs.

## Solution

Calculate the scanName using ClientID, the current workspace
configuration (project root(s) or active file), and code analysis scope.