fix(amazonq): context fixes for multiroot workspace setup (#6707)
## Problem
In multiroot workspace setup, the contextList UI displays files that are
deduped for the same files and doesn't resolve relative paths correctly.

1. In a multi-root workspace, the index is built on the root of the 1st
project after sorting workspaceFolders, thus the chunk response also
contains relativePath to the 1st project root after sorting. In order to
get the same relativePath from explicit @file, we also need to use the
root of the 1st project after sorting workspaceFolders.
2. Use a relativePath to projectRoot map to store the project root for
each relativePath so that on clicking the file(which only returns
relativePath in the event) we correctly construct the absolute path of
the file to open in the editor.

## Solution


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.