feat(lambda): link for serverlessland, fixes #6687
## Problem
This PR adds the functionality to open pattern in serverless-land
website. It addresses an issue with the `metadata.json` file path for
the extension by specifying the path after the build step in the
`copyFile.ts` file.

## Solution
Changes:
1. Removed the webview functionality and instead open the browser to
preview the patterns.
2. Removed the Github button.
3. Added a link at the bottom of pattern quick pick to explore the
[serverless land
website](https://serverlessland.com/patterns?services=lambda).
4. Ensured the proper path for the metadata file is accessible and it
gets executed properly after the build.
5. Added another entry point at the AWS Explorer high-level, accessible
when clicking the three-dot menu.