telemetry(amazonq): update to new module metric name
Problem:

Updated the telemetry types and had to change an existing name to match the new one.

Solution:

use `toolkit_didLoadModule` to indicate that the module successfully opened. In certain cases we will want to use `toolkit_willOpenModule` to indicate the start of it, but because it doesn't look like that was used already we can ignore it for now and revisit later.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>