telemetry(amazonq): update to new module metric name (#6717)
## Problem:

Updated the telemetry types and had to change an existing name to match
the new one.

## Solution:

use `toolkit_didLoadModule` to indicate that the module successfully
opened. In certain cases we will want to use `toolkit_willOpenModule` to
indicate the start of it, but because it doesn't look like that was used
already we can ignore it for now and revisit later.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>