feat(lambda): SAM local invoke can accept a runtime parameter #6724
## Problem
Customers are not able to local invoke with the runtime selected in the
webview UI. Currently the runtime used is the one specified on the
template. The feature was already added to SAMCLI
https://github.com/aws/aws-sam-cli/pull/7885.

## Solution
Allows customers to change runtimes when locally invoking in the local
invoke webview, without having to change it in their template and
rebuilding. Under the hood, the newly added ```--runtime``` option will
be appended to the ```samcli``` command.