telemetry(amazonq): /dev error classification #6772
## Problem

Today, telemetry emitted from the feature development is classified as
error, fault, or LLM-based failure across various parts of the code
base. This makes it difficult to know all the error types that exist and
how to classify them appropriately when emitting metrics.

## Solution
The solution here allows different exceptions to raise from different
parts of the code base and uses inheritance to model each error as a
client error, service error, or LLM-based error. When emitting
telemetry, we now only need to check whether the error received extends
from one of these 3 to know how to classify the telemetry.