refactor(workspaceutil): unify file collecting utilities (#6804)
## Problem
`collectFiles` and `collectFilesForIndex` serve different purposes, but
implement the same core functionality. This core functionality should be
not duplicated in two places.

## Solution
- Increase testing coverage for both methods. 
- Refactor `collectFiles` to be general enough to handle
`collectFilesForIndex` use case. This includes the ability to avoid
reading the files when we are indexing.
- `collectFilesForIndex` now directly calls `collectIndex` then does its
additional logic with the result.

## Verification
- Tested `@workspace` in chat with some different prompts. Seems to
behave the same as before.
- Tested `/review` on a project. 
---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.