refactor(auth): dont manual refresh auth webview on hide (#6836)
## Problem:

We have an explicit force refresh of the auth webview when it is
shown/hidden. There is not an obvious reason why we do this.

This causes problems since we expect that show/hiding of the webview
does not trigger any functional changes since we have configured the
webview to continue running even while hidden. This means that our
handlers for auth changes are still running even when auth is hidden.

For my specific issue, by force reloading the webview on every show/hide
it triggered the webview loading code each time, causing webview
telemetry events to be triggered more than expected.

## Solution:

Remove the force reload line. With this change we can move the edge case
handling code which throttled the telemetry emitted when it was
triggered on every hide/show.

I've done manual testing and have not noted any obvious regressions to
the expected user flow.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>