fix(appbuilder): Prevent parallel build processes on the same template (#6172)
## Context
We currently allow multiple build processes to run at the same time for
the same template.

## Problem

When running a build, it’s possible to start a second build for the same
project. This results in both builds failings

## Solution
Prevent parallel builds on the same template

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.