feat(amazonq lsp): add authFollowUpClicked implementation (#6886)
## Problem
There was no implementation for the `authFollowUpClicked` message
received from LSP for Amazon Q. Because of that, the `authenticate`
button that appears in case of expired SSO connection had no effect.

## Solution
I added the implementation as follows:
* If the received `authFollowUpType` is `'re-auth'` or
`'missing_scopes'`, initiate a re-authentication
* If the received `authFollowUpType` is `'full-auth'` or
`'use-supported-auth'`, delete the connection so a new full
authentication flow kicks off

I added a unit test file for this message type, which is extendable for
every command handled in `registerLanguageServerEventListener`

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.